refactor :  Calculatorable 와 Calculator 를 병합
* 계산기 버튼위의 숫자외에도 실수는 postfix에 입력되도록 변경
* 분할되어 있는 ErrorCase 병합
* .signConverter 에 대한 기능 추가
* StoryBoard에 tag를 통한 사용자 입력받음
